# vtc-service/admin-ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Verifiable Trust Community (VTC) API",
    "description": "Community lifecycle, ACL, audit, policy, credentials, endorsements, and cross-community recognition REST surface of a Verifiable Trust Community.",
    "contact": {
      "name": "Glenn Gore",
      "email": "glenn@affinidi.com"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.11.58"
  },
  "paths": {
    "/v1/admin/passkeys": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "adminPasskeyList",
        "responses": {
          "200": {
            "description": "Registered admin passkeys; empty when the                                       caller has enrolled none",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "403": {
            "description": "Caller is not an admin"
          }
        },
        "security": [
          {
            "bearer_jwt": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PasskeyListResponse": {
        "type": "object",
        "required": [
          "credentials"
        ],
        "properties": {
          "credentials": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegisteredCredential"
            },
            "description": "`credentials`, the name `auth/passkey/list/0.1` publishes. It was\n`passkeys` until #1112 — the same object under a name the schema does\nnot define, so no conforming client could find it."
          }
        }
      },
...
      "RegisteredCredential": {
        "type": "object",
        "description": "The wire view of a passkey, as `auth/passkey/list/0.1` publishes it.\n\nSeparate from [`RegisteredPasskey`] because that type is a **storage row**:\nit is `Deserialize`d out of the `admin:<did>` record, so its member names\nare the on-disk format and renaming them would orphan every passkey already\nenrolled. The two shapes disagree in exactly the two ways the schema cares\nabout, and both differences are the storage row's to keep:\n\n- `label` is published as `deviceLabel`, and is **absent** rather than empty\n  when nobody chose one. The schema is explicit that a consumer must not\n  synthesize a label, \"because an invented one is indistinguishable from a\n  chosen one to somebody deciding which credential to revoke\" — and `\"\"` is\n  an invented one.\n- `lastUsedAt` is omitted when unset. Sending `null` is a type error against\n  a `date-time` member, and the schema notes it deliberately does not\n  distinguish \"never used\" from \"not tracked\", since no binding could carry\n  the difference anyway.",
        "required": [
          "credentialId",
          "transports",
          "registeredAt"
        ],
        "properties": {
          "credentialId": {
            "type": "string"
          },
          "deviceLabel": {
            "type": [
              "string",
              "null"
            ]
          },
          "lastUsedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "registeredAt": {
            "type": "string",
            "format": "date-time"
          },
          "transports": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}