# vtc-service/src/routes/install.rs
//! `POST /v1/install/claim/{start,finish}` — WebAuthn install
//! ceremony for the very first admin.
//!
//! Implements **M0.5.2** of the VTC MVP Phase 0 plan. The flow:
//!
//! ```text
//! operator  ──install_token──▶  start  ──registration_id + ccr──▶ operator
//! operator  ──finish(webauthn_response)──▶  finish
//! finish    ──admin_did + setup_session_token──▶  operator
//! ```
//!
//! - `start` verifies the install token, takes the install-keyspace
//!   ceremony lock (`InstallTokenStore::start_claim`), and returns a
//!   WebAuthn `CreationChallengeResponse` constrained to Ed25519 via
//!   `vtc_service::webauthn::start_passkey_registration`, which
//!   accepts any algorithm the authenticator can produce
//!   (ES256, RS256, EdDSA).
//! - `finish` verifies the WebAuthn response, derives the candidate
//!   admin `did:key` from the credential's Ed25519 public key,
//!   consumes the install token, persists the passkey, and mints a
//!   short-lived setup-session token consumed by M0.6's
//!   `/v1/admin/bootstrap`.
//!
//! The WebAuthn attestation already proves the operator controls the
//! Ed25519 keypair that materialises the candidate did:key — modelled
//! on `affinidi-webvh-service`'s `enroll_finish` (no extra raw-bytes
//! binding signature). The previous design required a raw Ed25519
//! signature over a server challenge, which is impossible to produce
//! in a real browser (WebAuthn never exposes the private key).
//!
//! That reasoning lived only here until 2026-08. The published task
//! kept requiring the binding for two months after it was removed, and
//! the conformance sweep then read the gap as this service being behind
//! the spec on a security control — the opposite of what happened.
//! `install/claim/{start,finish}/0.2`
//! (trustoverip/dtgwg-trust-tasks-tf#263) drops it and carries the
//! argument in the specification itself. Claiming admin under a DID the
//! operator *already* controls is a separate, genuinely non-redundant
//! proof — ownership of a verification method in that DID's **active**
//! document, resolved live and signed outside the WebAuthn ceremony —
//! and is named there as a future version rather than approximated.
//!
//! Per-row state machine is the only gate on claim. A second
//! `start` on the same token after a successful `finish` returns
//! 401 because the row is now `Consumed`. The earlier global
//! "carve-out" lockdown is gone — each invite carries its own
//! Argon2id-hashed claim secret which `claim_start` verifies
//! before issuing the WebAuthn challenge.

use std::sync::Arc;

use axum::http::StatusCode;
use serde::{Deserialize, Serialize};
use tracing::info;
use uuid::Uuid;
use vti_common::auth::passkey::store::{
    PasskeyUser, delete_registration_user, get_registration_user, store_credential_mapping,
    store_passkey_user, store_registration_state, store_registration_user, take_registration_state,
};
use vti_common::error::AppError;
use webauthn_rs::prelude::{CreationChallengeResponse, RegisterPublicKeyCredential, Webauthn};

use crate::acl::admin::{AdminEntry, RegisteredPasskey, get_admin_entry, store_admin_entry};
use crate::error::TaskError;
use crate::install::{
    INSTALL_SESSION_DEFAULT_TTL_SECS, InstallTokenSigner, InstallTokenState, claim_secret,
    mint_install_session_token, parse_install_token,
};
use crate::server::AppState;

// ---------------------------------------------------------------------------
// Wire shapes
// ---------------------------------------------------------------------------

#[derive(Deserialize, utoipa::ToSchema)]
#[serde(rename_all = "camelCase")]
pub struct ClaimStartRequest {
    pub install_token: String,
    /// Out-of-band claim code the operator received alongside the
    /// install URL. Required when the persisted token row has a
    /// `claim_secret_hash`; ignored otherwise so legacy tokens
    /// (and tests) keep working. The route handler verifies the
    /// code against the hash before issuing the WebAuthn challenge
    /// — a wrong or missing code returns 401 with discriminated
    /// error codes (`claim_secret_required` / `claim_secret_invalid`).
    #[serde(default)]
    pub claim_secret: Option<String>,
}

/// Written by hand so the install token and claim code never reach a log: a derived `Debug` would print them.
impl std::fmt::Debug for ClaimStartRequest {
    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
        f.debug_struct("ClaimStartRequest")
            .field("install_token", &"<redacted>")
            .field(
                "claim_secret",
                &self.claim_secret.as_ref().map(|_| "<redacted>"),
            )
            .finish()
    }
}

#[derive(Debug, Serialize)]
#[serde(rename_all = "camelCase")]
#[derive(utoipa::ToSchema)]
pub struct ClaimStartResponse {
    /// Echoes the install token's `jti`. Consumer must pass this
    /// back to `claim/finish` so the server can index the stored
    /// registration state.
    pub registration_id: String,
    /// The WebAuthn `PublicKeyCredentialCreationOptions` payload —
    /// the operator's UA passes this to `navigator.credentials.create()`.
    #[schema(value_type = Object)]
    pub options: CreationChallengeResponse,
}

#[derive(Deserialize, utoipa::ToSchema)]
#[serde(rename_all = "camelCase")]
pub struct ClaimFinishRequest {
    pub install_token: String,
    pub registration_id: String,
    #[schema(value_type = Object)]
    pub webauthn_response: RegisterPublicKeyCredential,
}

/// Written by hand so the install token never reaches a log: a derived `Debug` would print it.
impl std::fmt::Debug for ClaimFinishRequest {
    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
        f.debug_struct("ClaimFinishRequest")
            .field("install_token", &"<redacted>")
            .field("registration_id", &self.registration_id)
            .field("webauthn_response", &self.webauthn_response)
            .finish()
    }
}

#[derive(Serialize)]
#[serde(rename_all = "camelCase")]
#[derive(utoipa::ToSchema)]
pub struct ClaimFinishResponse {
    pub admin_did: String,
    pub setup_session_token: String,
}

/// Written by hand so the setup-session token never reaches a log: a derived `Debug` would print it.
impl std::fmt::Debug for ClaimFinishResponse {
    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
        f.debug_struct("ClaimFinishResponse")
            .field("admin_did", &self.admin_did)
            .field("setup_session_token", &"<redacted>")
            .finish()
    }
}

// ---------------------------------------------------------------------------
// Declared error codes (#1600)
// ---------------------------------------------------------------------------

use trust_tasks_rs::specs::vtc::install::claim as claim_spec;

/// `vtc/install/claim/start:invalidToken` — missing, malformed, expired, or
/// already consumed.
pub const START_ERR_INVALID_TOKEN: &str = claim_spec::start::v0_2::error_codes::INVALID_TOKEN.code;
/// `vtc/install/claim/finish:invalidToken` — as for `start`.
pub const FINISH_ERR_INVALID_TOKEN: &str =
    claim_spec::finish::v0_2::error_codes::INVALID_TOKEN.code;
/// `vtc/install/claim/finish:registrationMismatch` — the `registrationId`
/// names no open enrolment for this token.
pub const FINISH_ERR_REGISTRATION_MISMATCH: &str =
    claim_spec::finish::v0_2::error_codes::REGISTRATION_MISMATCH.code;
/// `vtc/install/claim/finish:bindingInvalid` — the WebAuthn attestation
/// failed verification.
pub const FINISH_ERR_BINDING_INVALID: &str =
    claim_spec::finish::v0_2::error_codes::BINDING_INVALID.code;

/// Declare `code` on the token-state refusals: every `Unauthorized` the token
/// parser and the install state machine answer is the token being missing,
/// malformed, expired or consumed. A `Conflict` (a concurrent ceremony's
/// lock) and anything else stay undeclared.
fn token_error(code: &'static str) -> impl FnOnce(AppError) -> TaskError {
    move |e| match e {
        e @ AppError::Unauthorized(_) => TaskError::declared(code, e),
        e => TaskError::App(e),
    }
}

// ---------------------------------------------------------------------------
// Handlers
// ---------------------------------------------------------------------------

/// `vtc/install/claim/start/0.2` — begin the WebAuthn install ceremony for
/// the first admin. Pre-session: the install token is the credential. Called
/// from `trust_tasks::install_tasks`; no REST route mounts it.
pub(crate) async fn claim_start(
    state: &AppState,
    req: ClaimStartRequest,
) -> Result<ClaimStartResponse, TaskError> {
    let signer = require_install_signer(state)?;
    let webauthn = require_webauthn(state)?;
    let store = &state.install_store;

    let claims = parse_install_token(signer, &req.install_token)
        .map_err(token_error(START_ERR_INVALID_TOKEN))?;
    let jti = parse_jti(&claims.jti).map_err(token_error(START_ERR_INVALID_TOKEN))?;

    // Per-invite claim secret, verified BEFORE the ceremony lock (P0.21).
    // `start_claim` stamps `claimed_at`, a 300 s concurrency lock; if we
    // verified the secret *after* it, anyone holding only the install URL
    // could grief the legitimate operator by POSTing a wrong code every 5
    // minutes — each attempt would re-arm the lock, then get rejected. So
    // we peek the stored hash (no mutation) and verify first; only a
    // correct (or absent) secret reaches `start_claim`. Discriminated 401
    // codes let the browser surface the right hint (missing vs wrong code).
    // Rows without a hash skip this — covers legacy rows and tests. The
    // authoritative not-found / consumed / expired / concurrency checks
    // stay in `start_claim` below.
    if let Some(stored_hash) = store.peek_secret_hash(&jti).await? {
        let Some(supplied) = req.claim_secret.as_deref() else {
            return Err(TaskError::declared(
                START_ERR_INVALID_TOKEN,
                AppError::Unauthorized(
                    "claim_secret_required: this install URL needs the out-of-band claim code"
                        .into(),
                ),
            ));
        };
        // Argon2id verification is CPU-bound (~50–200 ms); run it on the
        // blocking pool so it doesn't stall the async REST runtime — this is
        // an *unauthenticated* route, so a few concurrent claim-starts must
        // not wedge every other request (P0.10).
        let supplied = supplied.to_string();
        let verified =
            tokio::task::spawn_blocking(move || claim_secret::verify(&supplied, &stored_hash))
                .await
                .map_err(|e| {
                    AppError::Internal(format!("claim-secret verify task failed: {e}"))
                })??;
        if !verified {
            return Err(TaskError::declared(
                START_ERR_INVALID_TOKEN,
                AppError::Unauthorized(
                    "claim_secret_invalid: the supplied claim code does not match".into(),
                ),
            ));
        }
    }

    // Take the ceremony lock. `start_claim` validates `Issued`, not
    // expired; on success the `claimed_at` window is set to "now"
    // so a second concurrent start sees the lock.
    store
        .start_claim(&jti)
        .await
        .map_err(token_error(START_ERR_INVALID_TOKEN))?;

    let user_uuid = jti;
    // Show the operator their admin DID in the authenticator's UI —
    // it's the identity the passkey will authenticate as. Modelled on
    // `affinidi-webvh-service`'s enroll-start which uses `user.did`
    // for both user_name and display_name. No algorithm restriction
    // here so any platform authenticator (Apple iCloud Keychain,
    // Windows Hello, Chrome passkeys, hardware keys) works — the
    // admin DID's shape is fixed in the install token, not derived
    // from the passkey.
    let (ccr, reg_state) = crate::webauthn::start_passkey_registration(
        webauthn,
        user_uuid,
        &claims.admin_did,
        &claims.admin_did,
        None,
    )?;

    // Persist the registration state under `jti` so `claim_finish`
    // can complete the ceremony against the same challenge.
    store_registration_state(&state.passkey_ks, &jti.to_string(), &reg_state).await?;

    // Carry the user UUID forward so the M0.6 bootstrap can look up
    // the PasskeyUser by registration_id without re-deriving it.
    store_registration_user(&state.passkey_ks, &jti.to_string(), &user_uuid).await?;

    info!(jti = %jti, "install claim ceremony started");

    Ok(ClaimStartResponse {
        registration_id: jti.to_string(),
        options: ccr,
    })
}

/// `vtc/install/claim/finish/0.2` — complete the WebAuthn install ceremony,
/// mint the admin DID + setup-session token. Pre-session. Called from
/// `trust_tasks::install_tasks`; no REST route mounts it.
pub(crate) async fn claim_finish(
    state: &AppState,
    req: ClaimFinishRequest,
) -> Result<ClaimFinishResponse, TaskError> {
    let signer = require_install_signer(state)?;
    let webauthn = require_webauthn(state)?;
    let store = &state.install_store;

    let claims = parse_install_token(signer, &req.install_token)
        .map_err(token_error(FINISH_ERR_INVALID_TOKEN))?;
    let jti = parse_jti(&claims.jti).map_err(token_error(FINISH_ERR_INVALID_TOKEN))?;
    // A `registrationId` that is not a UUID, or names another token's
    // enrolment, matches no open enrolment for this one.
    let registration_mismatch = || {
        TaskError::declared(
            FINISH_ERR_REGISTRATION_MISMATCH,
            AppError::Unauthorized("registration_id does not match install token".into()),
        )
    };
    let reg_id = Uuid::parse_str(&req.registration_id).map_err(|_| registration_mismatch())?;
    if reg_id != jti {
        return Err(registration_mismatch());
    }

    // Idempotent retry. Consume-first (below) is the security-correct
    // direction, but it leaves an availability sharp edge: a crash —
    // or a dropped response — between consuming the token and returning
    // permanently spends it, stranding the operator without the
    // `setup_session_token` `/admin/bootstrap` needs and with no way to
    // re-run the WebAuthn ceremony (its registration state was already
    // taken). So if the token is already `Consumed` *and* the ceremony
    // actually persisted an admin record, re-derive + re-mint the
    // setup-session token instead of hard-rejecting. The signed install
    // token's own `exp` (validated by `parse_install_token` above) bounds
    // the retry window; `start→finish→start` still rejects the second
    // `start` (it requires `Issued`).
    if let Some(InstallTokenState::Consumed { admin_did, .. }) = store.get_token(&jti).await? {
        let admin_did = admin_did.unwrap_or_else(|| claims.admin_did.clone());
        // Only re-issue if the first call got far enough to persist the
        // admin — otherwise there's nothing to bootstrap against.
        if get_admin_entry(&state.passkey_ks, &admin_did)
            .await?
            .is_some()
        {
            info!(jti = %jti, %admin_did, "install claim finish replayed; re-issuing setup token");
            return Ok(issue_setup_session(state, signer, admin_did, &jti).await?);
        }
        return Err(TaskError::declared(
            FINISH_ERR_INVALID_TOKEN,
            AppError::Unauthorized("install token consumed".into()),
        ));
    }

    let reg_state = take_registration_state(&state.passkey_ks, &jti.to_string())
        .await?
        .ok_or_else(|| {
            TaskError::declared(
                FINISH_ERR_REGISTRATION_MISMATCH,
                AppError::Unauthorized(
                    "no registration in progress for this install token (start the ceremony \
                     first)"
                        .into(),
                ),
            )
        })?;

    // Run the WebAuthn ceremony. Any algorithm the authenticator
    // offers is acceptable — the admin DID is carried in the install
    // token, not derived from the passkey, so a platform passkey
    // (Apple iCloud Keychain, Windows Hello, Chrome passkeys, etc.)
    // works regardless of whether it produces ES256, EdDSA, or RS256.
    let passkey = webauthn
        .finish_passkey_registration(&req.webauthn_response, &reg_state)
        .map_err(|e| {
            TaskError::declared(
                FINISH_ERR_BINDING_INVALID,
                AppError::Authentication(format!("passkey registration failed: {e}")),
            )
        })?;
    let admin_did = claims.admin_did.clone();

    // Consume the install token (Issued → Consumed). Carve-out stays
    // open until M0.6's bootstrap closes it.
    store
        .finish_claim(&jti)
        .await
        .map_err(token_error(FINISH_ERR_INVALID_TOKEN))?;

    // Persist the passkey + credential mapping so M0.6's bootstrap
    // and subsequent passkey login can find the credential.
    let user_uuid = get_registration_user(&state.passkey_ks, &jti.to_string())
        .await?
        .ok_or_else(|| AppError::Internal("missing registration_user mapping".into()))?;
    delete_registration_user(&state.passkey_ks, &jti.to_string()).await?;
    let user = PasskeyUser {
        user_uuid,
        did: admin_did.clone(),
        display_name: admin_did.clone(),
        credentials: vec![passkey.clone()],
    };
    store_passkey_user(&state.passkey_ks, &user).await?;
    let cred_id_hex = hex::encode(passkey.cred_id().as_ref() as &[u8]);
    store_credential_mapping(&state.passkey_ks, &cred_id_hex, user_uuid).await?;

    // Create or append to the AdminEntry so `GET /v1/admin/passkeys`
    // can list the just-registered device. The first-admin install
    // flow then takes the same AdminEntry through `/v1/admin/bootstrap`;
    // the `vtc admin invite` flow never reaches bootstrap (carve-out
    // already closed by the first admin), so this write is the only
    // place the AdminEntry is populated for invited admins.
    let now = chrono::Utc::now();
    let registered = RegisteredPasskey {
        credential_id: cred_id_hex.clone(),
        label: "install".into(),
        transports: Vec::new(),
        registered_at: now,
        last_used_at: None,
    };
    let admin_entry = match get_admin_entry(&state.passkey_ks, &admin_did).await? {
        Some(mut existing) => {
            // Dedupe by credential_id — a retry of the same ceremony
            // must not double-list the device.
            if !existing
                .passkeys
                .iter()
                .any(|p| p.credential_id == cred_id_hex)
            {
                existing.passkeys.push(registered);
            }
            existing
        }
        None => AdminEntry {
            did: admin_did.clone(),
            passkeys: vec![registered],
            extensions: serde_json::Value::Null,
            created_at: now,
        },
    };
    store_admin_entry(&state.passkey_ks, &admin_entry).await?;

    // Audit the first-admin passkey enrolment — the install-claim flow is the
    // only place an invited admin's passkey is registered, so without this the
    // device that gained admin authority leaves no audit trail.
    if let Some(writer) = state.audit_writer.as_ref() {
        writer
            .write(
                &admin_did,
                Some(&admin_did),
                vti_common::audit::AuditEvent::AdminPasskeyRegistered(
                    vti_common::audit::AdminPasskeyData {
                        credential_id_hex: cred_id_hex.clone(),
                        label: "install".into(),
                        transports: Vec::new(),
                    },
                ),
            )
            .await?;
    }

    info!(jti = %jti, %admin_did, "install claim ceremony completed");

    Ok(issue_setup_session(state, signer, admin_did, &jti).await?)
}

/// Mint the `setup_session_token` for `admin_did` + build the
/// `claim/finish` response. Shared by the first-completion path and the
/// idempotent-replay path so a retry returns a usable token too.
async fn issue_setup_session(
    state: &AppState,
    signer: &InstallTokenSigner,
    admin_did: String,
    jti: &Uuid,
) -> Result<ClaimFinishResponse, AppError> {
    let issuer_did = state
        .config
        .read()
        .await
        .vtc_did
        .clone()
        .unwrap_or_else(|| "did:key:vtc-install-uninitialised".to_string());

    let setup_session_token = mint_install_session_token(
        signer,
        &issuer_did,
        &admin_did,
        &jti.to_string(),
        INSTALL_SESSION_DEFAULT_TTL_SECS,
    )?;

    Ok(ClaimFinishResponse {
        admin_did,
        setup_session_token,
    })
}

// ---------------------------------------------------------------------------
// Helpers
// ---------------------------------------------------------------------------

fn require_install_signer(state: &AppState) -> Result<&Arc<InstallTokenSigner>, AppError> {
    state
        .install_signer
        .as_ref()
        .ok_or_else(|| AppError::ServiceError {
            status: StatusCode::SERVICE_UNAVAILABLE,
            message: "install signer not configured (run setup first)".into(),
        })
}

fn require_webauthn(state: &AppState) -> Result<&Webauthn, AppError> {
    state
        .webauthn
        .as_deref()
        .ok_or_else(|| AppError::ServiceError {
            status: StatusCode::SERVICE_UNAVAILABLE,
            message: "WebAuthn not configured (public_url required)".into(),
        })
}

fn parse_jti(s: &str) -> Result<Uuid, AppError> {
    Uuid::parse_str(s)
        .map_err(|_| AppError::Unauthorized("invalid install token (malformed jti)".into()))
}

// All Ed25519-derivation helpers (extract_ed25519_public_key,
// walk_eddsa_x, decode_x_value, ed25519_pub_to_did_key) lived here
// to project the WebAuthn key into the admin did:key. The redesigned
// flow carries admin_did in the install token, so the helpers are
// no longer needed. Removing them lets the install ceremony accept
// any algorithm the authenticator offers (ES256 from platform
// passkeys, EdDSA from hardware keys, etc.).